// src/uiLibrary/components-ui/styles/models/IStyleConfig.ts
import EnumThemeVariant from "../../enums/EnumThemeVariant";
import IButtonConfig from "./sections/IButtonConfig";
import ISolidColorConfig from "./sections/ISolidColorConfig";
import ITextFieldConfig from "./sections/ITextFieldConfig";

export default interface IStyleConfig {
  theme: EnumThemeVariant;

  // Generic - Button

  textField: ITextFieldConfig;
  button: IButtonConfig;
  solid: ISolidColorConfig;
}